requirements.txt
argparse==1.4.0
pillow==8.3.1
